+++ soffa-foundation-api/src/main/java/io/soffa/foundation/core/models/Validatable.java
@@ -1,7 +1,0 @@
-package io.soffa.foundation.core.models;
-
-public interface Validatable {
-
-    void validate();
-
-}