+++ soffa-foundation-api/src/test/java/io/soffa/foundation/model/User.java
@@ -1,9 +1,0 @@
-package io.soffa.foundation.model;
-
-import lombok.Data;
-
-@Data
-public class User {
-
-    private String username;
-}