+++ soffa-foundation-application/src/main/java/io/soffa/foundation/core/Constants.java
@@ -1,5 +1,0 @@
-package io.soffa.foundation.core;
-
-public interface Constants {
-
-}