+++ soffa-foundation-application/src/main/java/io/soffa/foundation/core/events/DatabaseReadyEvent.java
@@ -1,5 +1,0 @@
-package io.soffa.foundation.core.events;
-
-
-public class DatabaseReadyEvent {
-}