+++ soffa-foundation-application/src/main/java/io/soffa/foundation/core/events/OnServiceStarted.java
@@ -1,6 +1,0 @@
-package io.soffa.foundation.core.events;
-
-import io.soffa.foundation.core.Operation;
-
-public interface OnServiceStarted extends Operation<ServiceInfo, Void> {
-}