+++ soffa-foundation-application/src/main/java/io/soffa/foundation/models/Frequency.java
@@ -1,9 +1,0 @@
-package io.soffa.foundation.models;
-
-public enum Frequency {
-    DAY,
-    WEEK,
-    MONTH,
-    YEAR,
-    NONE
-}