+++ soffa-foundation-application/src/main/java/io/soffa/foundation/models/TokenType.java
@@ -1,6 +1,0 @@
-package io.soffa.foundation.models;
-
-public enum TokenType {
-    JWT,
-    PASETO
-}