+++ soffa-foundation-application/src/main/java/io/soffa/foundation/models/files/AttachmentType.java
@@ -1,9 +1,0 @@
-package io.soffa.foundation.models.files;
-
-public enum AttachmentType {
-    IMAGE,
-    VIDEO,
-    PDF,
-    AUDIO,
-    OTHER
-}