+++ soffa-foundation-commons/src/main/java/io/soffa/foundation/commons/PasswordEncoder.java
@@ -1,7 +1,0 @@
-package io.soffa.foundation.commons;
-
-public interface PasswordEncoder {
-    String encode(String rawPassword);
-
-    boolean matches(String rawPassword, String encryptedPassword);
-}