+++ soffa-foundation-commons/src/main/java/io/soffa/foundation/errors/ManagedException.java
@@ -1,4 +1,0 @@
-package io.soffa.foundation.errors;
-
-public interface ManagedException   {
-}