+++ soffa-foundation-service-test/src/main/resources/karate/basic-auth.js
@@ -1,4 +1,0 @@
-function fn(creds) {
-    var helper = Java.type('io.soffa.foundation.test.KarateTestUtil');
-    return helper.basicAuth(creds.username, creds.password);
-}