+++ soffa-foundation-service/src/main/java/io/soffa/foundation/service/core/config/NoopAuthManager.java
@@ -1,7 +1,0 @@
-package io.soffa.foundation.service.core.config;
-
-import io.soffa.foundation.core.security.AuthManager;
-
-public class NoopAuthManager implements AuthManager {
-
-}